refactor: 도메인 데이터 파일 분리
